Replace main loop polling with events, defer module switching
Replaces the old keyboard and mouse polling loop with manual ticks with
code that uses the allegro5 events system.

Defer switching between modules by calling LoadModule as part of the
main event loop. This should ensure that the current module isn't closed
while it is processing a key, mouse, or broadcasted event.