Replace the Flux and TileScroller implementations
= Flux

Replace Flux with two classes: Flux and FluxInfo. Flux is a data
object that loads the flux coordinates from the galaxy.xml file.
FluxInfo is stored in the GameState and is information about
what the user has encountered and mapped.

Remove the random flux generation when ModuleInterstellarTravel is
loaded, replace with a static list loaded from galaxy.xml.o

Add bi-directional travel using fluxes. When exiting a flux, the
ship is placed at a random tile adjacent to the other flux exit.

Flux entry and exits are drawn on the ModuleStarmap and
ModuleInterstellarTravel if the the navigator has enough skill. If
the entry is visible and the navigator is skilled, the path is also
drawn on the starmap.

= TileScroller

This is a major rework of the TileScroller class to better
match the way it was being used.

The motivation for this was to move the coordinate transformations
between pixel and tile coordinates to the internals of the class instead
of having to do the adjustments each time a tile value is set or the
scroll position is moved. This was complicated and not being done
accurately in all places. It also was impossible to scroll to the
edges of the galaxy without problems.

I also noticed that there were many places in the code using the tile
scroller as a way of determining information about the system or galaxy.
The data already exists in the DataMgr or GameState in a sensible form,
so there's no real need to try to pull it out of the view of the data.
I replaced those tile lookups and removed the method that allowed
reading a tile.

From a usage perspective, the major changes are:

There's a new TileSet class that defines the parameters of the tile
image (bitmap, dimensions, layout) separate from the TileScroller to
make the constructor a bit simpler.

The TileScroller constructor changed parameters a bit to allow callers
to completely configure the object at creation time.

The TileScroller automatically updates the scroll buffer when it draws
if a tile changed or the scroll position moved to a different starting
tile. The method to update the scroll buffer is made private.

Replace the names of the TileScroll from lowerCamelCase to snake_case.