Fix braking drift
When speed is close to 0, the code was adding or subtracting a braking
offset to the speed which was greater than the threshold for when to
stop. This caused the ship to drift about forward and backward. This
patch makes the stop threshold match the breaking adjustment, so that
the ship actually stops.