Rework Starmap code, cleanups
This updates the Starmap code to remove the magic numbers from the class
implementation, to hopefully make it easier to see what's going on with
the coordinate systems for drawing and selecting things.

Also, move the fuel consumption calculation to the Ship object in the
GameState, rather than querying the sprite for info and then doing the
computation in the starmap module.